requirements.txt
flask
numpy
pandas
scikit-learn
scikit-image
opencv-python-headless
xgboost
pillow
joblib
scipy